Some simplifications
