+++ frontend/src/components/cars-categories/mock-brands.ts
@@ -1,0 +1,57 @@
+const mockBrands = [
+  {
+    'id': 90,
+    'name': 'BMW',
+    'autoria_code': 1,
+    'logo_url':
+      'https://upload.wikimedia.org/wikipedia/commons/thumb/4/44/BMW.svg/2048px-BMW.svg.png',
+  },
+  {
+    'id': 98,
+    'name': 'Audi',
+    'autoria_code': 2,
+    'logo_url': 'https://pngimg.com/uploads/car_logo/car_logo_PNG1640.png',
+  },
+  {
+    'id': 6,
+    'name': 'Mazda',
+    'autoria_code': 3,
+    'logo_url': 'https://www.pngall.com/wp-content/uploads/2/Mazda-PNG-Pic.png',
+  },
+  {
+    'id': 3,
+    'name': 'Chery',
+    'autoria_code': 4,
+    'logo_url':
+      'https://logos-world.net/wp-content/uploads/2021/08/Chery-Logo.png',
+  },
+  {
+    'id': 9,
+    'name': 'Volkswagen',
+    'autoria_code': 5,
+    'logo_url': 'https://logojinni.com/image/logos/volkswagen-3.svg',
+  },
+  {
+    'id': 5,
+    'name': 'Kia',
+    'autoria_code': 6,
+    'logo_url':
+      'https://static.wixstatic.com/media/f2bf43_bbd9119628ef4f54a9c502fb68090b59~mv2.png',
+  },
+  {
+    'id': 8,
+    'name': 'Subaru',
+    'autoria_code': 7,
+    'logo_url':
+      'https://images.vectorhq.com/images/previews/99e/subaru-logo-psd-462096.png',
+  },
+  {
+    'id': 94,
+    'name': 'Porsche',
+    'autoria_code': 8,
+    'logo_url':
+      'https://auto-shop.dp.ua/uploads/brand/7cfd7e04ff4cc69fe0e7c9513f5b4e48.png',
+  },
+];
+
+export { mockBrands };
