--- frontend/src/components/cars-categories/mockBrands.ts
@@ -1,71 +1,0 @@
-const mockBrands = [
-  {
-    'name': 'BMW',
-    'marka_id': 90,
-    'eng': 'bmw',
-    'country_id': 392,
-    'category_id': 1,
-    'logo':
-      'https://upload.wikimedia.org/wikipedia/commons/thumb/4/44/BMW.svg/2048px-BMW.svg.png',
-  },
-  {
-    'name': 'Audi',
-    'marka_id': 98,
-    'eng': 'acura',
-    'country_id': 392,
-    'category_id': 1,
-    'logo': 'https://pngimg.com/uploads/car_logo/car_logo_PNG1640.png',
-  },
-  {
-    'name': 'Mazda',
-    'marka_id': 6,
-    'eng': 'audi',
-    'country_id': 276,
-    'category_id': 1,
-    'logo': 'https://www.pngall.com/wp-content/uploads/2/Mazda-PNG-Pic.png',
-  },
-  {
-    'name': 'Chery',
-    'marka_id': 3,
-    'eng': 'alfa-romeo',
-    'country_id': 380,
-    'category_id': 1,
-    'logo': 'https://logos-world.net/wp-content/uploads/2021/08/Chery-Logo.png',
-  },
-  {
-    'name': 'Volkswagen',
-    'marka_id': 9,
-    'eng': 'bentley',
-    'country_id': 392,
-    'category_id': 1,
-    'logo': 'https://logojinni.com/image/logos/volkswagen-3.svg',
-  },
-  {
-    'name': 'Kia',
-    'marka_id': 5,
-    'eng': 'aston-martin',
-    'country_id': 826,
-    'category_id': 1,
-    'logo':
-      'https://static.wixstatic.com/media/f2bf43_bbd9119628ef4f54a9c502fb68090b59~mv2.png',
-  },
-  {
-    'name': 'Subaru',
-    'marka_id': 8,
-    'eng': 'bentley',
-    'country_id': 826,
-    'category_id': 1,
-    'logo':
-      'https://images.vectorhq.com/images/previews/99e/subaru-logo-psd-462096.png',
-  },
-  {
-    'name': 'Porsche',
-    'marka_id': 94,
-    'eng': 'porsche',
-    'country_id': 392,
-    'category_id': 1,
-    'logo': 'https://upload.wikimedia.org/wikipedia/uk/5/55/Porsche_logo.png',
-  },
-];
-
-export { mockBrands };
