+++ frontend/src/types/brand.type.ts
@@ -1,8 +1,6 @@
 export type Brand = {
+  id: number;
   name: string;
-  marka_id: number;
-  eng: string;
-  country_id: number;
-  category_id: number;
-  logo: string;
+  logo_url: string;
+  autoria_code: number;
 };
